Chrysler: Common safety checksum (#3231)
* Chrysler: Common safety checksum

* sanity check all counters are length 4
